+++ src/main/java/team/silvertown/masil/auth/service/OAuthService.java
@@ -1,4 +1,4 @@
-package team.silvertown.masil.user.service;
+package team.silvertown.masil.auth.service;
 
 import java.io.IOException;
 import team.silvertown.masil.user.dto.OAuthResponse;
