+++ src/main/java/team/silvertown/masil/config/security/WebSecurityConfig.java
@@ -9,7 +9,6 @@ import org.springframework.security.config.annotation.web.configurers.AbstractHt
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import team.silvertown.masil.config.jwt.JwtAuthenticationFilter;
 
 @Configuration
 @EnableWebSecurity
