+++ src/test/java/team/silvertown/masil/user/service/KakaoOAuthServiceTest.java
@@ -13,6 +13,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import team.silvertown.masil.auth.service.KakaoOAuthService;
 import team.silvertown.masil.security.exception.InvalidAuthenticationException;
 import team.silvertown.masil.user.dto.OAuthResponse;
 import team.silvertown.masil.user.exception.UserErrorCode;
